add more feautre: register (user), login(access_token), add contact(require auth, update contact_list of user), delete contact (require auth except admin), delete contact by id (require auth except admin), get contacts by userId, get favorite contacts by userId

+++ src/app/controllers/contact.controller.js
@@ -1,20 +1,25 @@
-const ApiError = require("../api-error");
-const ContactService = require("../services/contact.service");
-const MongoDB = require("../utils/mongodb.util");
+const ApiError = require('../api-error');
+const ContactService = require('../services/contact.service');
+const UserService = require('../services/user.service');
+const MongoDB = require('../utils/mongodb.util');
 
 // Create and Save a new Contact
 exports.create = async (req, res, next) => {
     if (!req.body?.name) {
-        return next(new ApiError(400, "Name cannot be empty"));
+        return next(new ApiError(400, 'Name cannot be empty'));
     }
 
     try {
         const contactService = new ContactService(MongoDB.client);
-        const document = await contactService.create(req.body);
-        return res.send(document);
+        const userService = new UserService(MongoDB.client);
+
+        const contact = await contactService.create(req.body, req.user._id);
+        await userService.addContact(contact._id, req.user._id);
+
+        return res.send(contact);
     } catch (error) {
         return next(
-            new ApiError(500, "An error occurred while creating the contact")
+            new ApiError(500, 'An error occurred while creating the contact')
         );
     }
 };
@@ -33,7 +38,7 @@ exports.findAll = async (req, res, next) => {
         }
     } catch (error) {
         return next(
-            new ApiError(500, "An error occurred while retrieving contacts")
+            new ApiError(500, 'An error occurred while retrieving contacts')
         );
     }
 
@@ -46,7 +51,7 @@ exports.findOne = async (req, res, next) => {
         const contactService = new ContactService(MongoDB.client);
         const document = await contactService.findById(req.params.id);
         if (!document) {
-            return next(new ApiError(404, "Contact not found"));
+            return next(new ApiError(404, 'Contact not found'));
         }
         return res.send(document);
     } catch (error) {
@@ -62,32 +67,63 @@ exports.findOne = async (req, res, next) => {
 // Update a contact by the id in the request
 exports.update = async (req, res, next) => {
     if (Object.keys(req.body).length === 0) {
-        return next(new ApiError(400, "Data to update cannot be empty"));
+        return next(new ApiError(400, 'Data to update cannot be empty'));
     }
 
     try {
         const contactService = new ContactService(MongoDB.client);
+
+        const contact = await contactService.findById(req.params.id);
+
+        if (contact.userId !== req.user._id && !req.user.isAdmin) {
+            return next(
+                new ApiError(
+                    403,
+                    "You don't have permission to perform this action"
+                )
+            );
+        }
+
         const document = await contactService.update(req.params.id, req.body);
         if (!document) {
-            return next(new ApiError(404, "Contact not found"));
+            return next(new ApiError(404, 'Contact not found'));
         }
-        return res.send({ message: "Contact was updated successfully" });
+        return res.send({ message: 'Contact was updated successfully' });
     } catch (error) {
         return next(
-            new ApiError(500, `Error updating contact with id=${req.params.id}`)
+            new ApiError(
+                500,
+                `Error updating contact with id = ${req.params.id}`
+            )
         );
     }
 };
 
 // Delete a contact with the specified id in the request
-exports.delete = async (req, res) => {
+exports.delete = async (req, res, next) => {
     try {
         const contactService = new ContactService(MongoDB.client);
+        const userService = new UserService(MongoDB.client);
+
+        const contact = await contactService.findById(req.params.id);
+
+        if (contact.userId !== req.user._id && !req.user.isAdmin) {
+            return next(
+                new ApiError(
+                    403,
+                    "You don't have permission to perform this action"
+                )
+            );
+        }
+
         const document = await contactService.delete(req.params.id);
         if (!document) {
-            return next(new ApiError(404, "Contact not found"));
+            return next(new ApiError(404, 'Contact not found'));
         }
-        return res.send({ message: "Contact was deleted successfully" });
+
+        await userService.removeContact(contact._id, req.user._id);
+
+        return res.send({ message: 'Contact was deleted successfully' });
     } catch (error) {
         return next(
             new ApiError(
@@ -108,7 +144,7 @@ exports.deleteAll = async (req, res, next) => {
         });
     } catch (error) {
         return next(
-            new ApiError(500, "An error occurred while removing all contacts")
+            new ApiError(500, 'An error occurred while removing all contacts')
         );
     }
 };
@@ -123,7 +159,7 @@ exports.findAllFavorite = async (req, res, next) => {
         return next(
             new ApiError(
                 500,
-                "An error occurred while retrieving favorite contacts"
+                'An error occurred while retrieving favorite contacts'
             )
         );
     }
